requirements.txt
python==3.6.4
pytorch==1.0.1
numpy==1.15.2
pandas==0.24.2
scipy==1.1.0
scikit-image==0.14.2
scikit-learn==0.20.3
